Merge to fips-20250107: delocate: remove the need to demangle local variables
We need to demangle because the backing storage for a delocated BSS
symbol is declared static. The compiler seems to be always give it a
C++-mangled name. But we relied on those names to infer the names of the
bss_get functions that the rest of the code wanted.

Instead, if we make the backing storage non-static, we get the name we
want. Ideally we would avoid having to extract those names (see
https://boringssl-review.googlesource.com/c/boringssl/+/70849), but
faking the linker script behavior in delocate seemed tricky, so I opted
to do this and then we can finish up the linker script version later.

While I'm here, prefix the underlying names with 'bcm_' so we're not
squatting as much of the namespace.

(cherry picked from commit 673e61fc215b178a90c0e67858bbf162c8158993)

Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/76847
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: Adam Langley <agl@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Change-Id: I79227c1fe4f1e5121ce170ab6937acad4a1537d0
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/76952
Reviewed-by: David Benjamin <davidben@google.com>